Added publish to artifactory to PluginPublishingPlugin
